awcy: Add VVC_VTM support
This is dependent on the rd_tool changeset for the same set of support

Right now, we have RA, LD, AI, and also GOP Parallel.
The QPs and cofigs are not fine-tuned but should be okay to have
JVET_CTC as it is first.

This also unbreaks the CTC_CSV Export but XLSM will be broken due to uneven QPs