Fix VP8/9 build configuration
Not sure how this ever worked--- but both the DDockerfile and the
libvpx build use the environment variable CONFIG_DIR for completely
different dirs.  This causes the libvpx builds to fail in Docker.

re-set CONFIG_DIR to what libvpx wants during its build.

A better fix would be to rename our variable to AWCY_CONFIG_DIR, but
that touches a lot more code.