Make 'perf' friendlier to non-Ubuntu Docker hosts
Debian's scripted perf wrapper does not work when run in Docker on a
non Debian/Ubuntu host.  Specifically it insists on a Debian kernel
and perf matching its own naming scheme, and the kernel is an
impossibility on eg RH.

We don't need kernel symbols, just the userland interface, so any
close 'linux-tools-*-generic' will work. Extend the script wrapper to
look for these close matches, and install it when building Dockerfile.worker