csv_export: Deprecate sorting for clipNames
CTC now reached a stage where the clipList is of different styles like,
a) CLIPNAME, b) ClipName, c) clipName, d) clipname, e) Clip_name, and
many other variants of stying for sequences name.
Handling all of them is now impossible without modifying clip names.
So let's trust and do the hardcoding of the clip in the exact order we want.